+++ src/main/java/spring/backend/activity/application/ReadMonthlyActivityOverviewService.java
@@ -5,9 +5,9 @@ import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import spring.backend.activity.domain.value.Keyword.Category;
-import spring.backend.activity.dto.request.MonthlyActivityOverviewRequest;
+import spring.backend.activity.presentation.dto.request.MonthlyActivityOverviewRequest;
 import spring.backend.activity.dto.response.MonthlyActivityCountByKeywordResponse;
-import spring.backend.activity.dto.response.MonthlyActivityOverviewResponse;
+import spring.backend.activity.presentation.dto.response.MonthlyActivityOverviewResponse;
 import spring.backend.activity.dto.response.MonthlySavedTimeAndActivityCountResponse;
 import spring.backend.activity.infrastructure.persistence.jpa.value.KeywordJpaValue;
 import spring.backend.activity.query.dao.ActivityDao;
