+++ src/main/java/spring/backend/activity/presentation/FinishActivityController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 import spring.backend.activity.application.FinishActivityService;
-import spring.backend.activity.dto.response.FinishActivityResponse;
+import spring.backend.activity.presentation.dto.response.FinishActivityResponse;
 import spring.backend.activity.presentation.swagger.FinishActivitySwagger;
 import spring.backend.core.configuration.argumentresolver.AuthorizedMember;
 import spring.backend.core.configuration.interceptor.Authorization;
