+++ src/main/java/spring/backend/activity/presentation/QuickStartActivitySelectController.java
@@ -8,11 +8,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import spring.backend.activity.application.QuickStartActivitySelectService;
-import spring.backend.activity.dto.request.QuickStartActivitySelectRequest;
-import spring.backend.activity.dto.response.QuickStartActivitySelectResponse;
+import spring.backend.activity.presentation.dto.request.QuickStartActivitySelectRequest;
+import spring.backend.activity.presentation.dto.response.QuickStartActivitySelectResponse;
 import spring.backend.activity.presentation.swagger.QuickStartActivitySelectSwagger;
 import spring.backend.core.configuration.argumentresolver.AuthorizedMember;
-import spring.backend.core.configuration.argumentresolver.LoginMember;
 import spring.backend.core.configuration.interceptor.Authorization;
 import spring.backend.core.presentation.RestResponse;
 import spring.backend.member.domain.entity.Member;
