+++ src/main/java/spring/backend/activity/presentation/ReadActivitiesByMemberAndKeywordInMonthController.java
@@ -3,12 +3,11 @@ package spring.backend.activity.presentation;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 import spring.backend.activity.application.ReadActivitiesByMemberAndKeywordInMonthService;
-import spring.backend.activity.dto.request.ActivitiesByMemberAndKeywordInMonthRequest;
-import spring.backend.activity.dto.response.ActivitiesByMemberAndKeywordInMonthResponse;
+import spring.backend.activity.presentation.dto.request.ActivitiesByMemberAndKeywordInMonthRequest;
+import spring.backend.activity.presentation.dto.response.ActivitiesByMemberAndKeywordInMonthResponse;
 import spring.backend.activity.presentation.swagger.ReadActivitiesByMemberAndKeywordInMonthSwagger;
 import spring.backend.core.configuration.argumentresolver.AuthorizedMember;
 import spring.backend.core.configuration.interceptor.Authorization;
