+++ src/main/java/spring/backend/activity/presentation/ReadActivityCalendarController.java
@@ -5,10 +5,10 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
-import spring.backend.activity.dto.request.ReadActivityCalendarRequest;
-import spring.backend.activity.dto.response.ActivityCalendarResponse;
-import spring.backend.activity.dto.response.UserMonthlyActivityDetail;
-import spring.backend.activity.dto.response.UserMonthlyActivitySummary;
+import spring.backend.activity.presentation.dto.request.ReadActivityCalendarRequest;
+import spring.backend.activity.presentation.dto.response.ActivityCalendarResponse;
+import spring.backend.activity.presentation.dto.response.UserMonthlyActivityDetail;
+import spring.backend.activity.presentation.dto.response.UserMonthlyActivitySummary;
 import spring.backend.activity.presentation.swagger.ReadActivityCalendarSwagger;
 import spring.backend.activity.query.dao.ActivityDao;
 import spring.backend.core.configuration.argumentresolver.AuthorizedMember;
