+++ src/main/java/spring/backend/activity/presentation/ReadMonthlyActivityOverviewController.java
@@ -4,11 +4,10 @@ import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RestController;
 import spring.backend.activity.application.ReadMonthlyActivityOverviewService;
-import spring.backend.activity.dto.request.MonthlyActivityOverviewRequest;
-import spring.backend.activity.dto.response.MonthlyActivityOverviewResponse;
+import spring.backend.activity.presentation.dto.request.MonthlyActivityOverviewRequest;
+import spring.backend.activity.presentation.dto.response.MonthlyActivityOverviewResponse;
 import spring.backend.activity.presentation.swagger.ReadMonthlyActivityOverviewSwagger;
 import spring.backend.core.configuration.argumentresolver.AuthorizedMember;
 import spring.backend.core.configuration.interceptor.Authorization;
