+++ src/main/java/spring/backend/activity/presentation/UserActivitySelectController.java
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import spring.backend.activity.application.UserActivitySelectService;
-import spring.backend.activity.dto.request.UserActivitySelectRequest;
-import spring.backend.activity.dto.response.UserActivitySelectResponse;
+import spring.backend.activity.presentation.dto.request.UserActivitySelectRequest;
+import spring.backend.activity.presentation.dto.response.UserActivitySelectResponse;
 import spring.backend.activity.presentation.swagger.UserActivitySelectSwagger;
 import spring.backend.core.configuration.argumentresolver.AuthorizedMember;
 import spring.backend.core.configuration.interceptor.Authorization;
