+++ src/main/java/spring/backend/activity/presentation/dto/request/ReadActivityCalendarRequest.java
@@ -1,4 +1,4 @@
-package spring.backend.activity.dto.request;
+package spring.backend.activity.presentation.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Max;
