+++ src/main/java/spring/backend/activity/presentation/dto/response/ActivityCalendarResponse.java
@@ -1,4 +1,4 @@
-package spring.backend.activity.dto.response;
+package spring.backend.activity.presentation.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
