+++ src/main/java/spring/backend/activity/presentation/dto/response/FinishActivityResponse.java
@@ -1,6 +1,7 @@
-package spring.backend.activity.dto.response;
+package spring.backend.activity.presentation.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
+import spring.backend.activity.dto.response.ActivityInfo;
 import spring.backend.member.dto.response.HomeMemberInfoResponse;
 
 public record FinishActivityResponse(
