+++ src/main/java/spring/backend/activity/query/dao/ActivityDao.java
@@ -3,8 +3,8 @@ package spring.backend.activity.query.dao;
 import spring.backend.activity.domain.value.Keyword;
 import spring.backend.activity.dto.response.*;
 import spring.backend.activity.dto.response.HomeActivityInfoResponse;
-import spring.backend.activity.dto.response.UserMonthlyActivityDetail;
-import spring.backend.activity.dto.response.UserMonthlyActivitySummary;
+import spring.backend.activity.presentation.dto.response.UserMonthlyActivityDetail;
+import spring.backend.activity.presentation.dto.response.UserMonthlyActivitySummary;
 
 import java.time.LocalDateTime;
 import java.util.List;
