+++ src/test/java/spring/backend/activity/application/UserActivitySelectServiceTest.java
@@ -11,8 +11,8 @@ import spring.backend.activity.domain.entity.Activity;
 import spring.backend.activity.domain.repository.ActivityRepository;
 import spring.backend.activity.domain.value.Keyword;
 import spring.backend.activity.domain.value.Type;
-import spring.backend.activity.dto.request.UserActivitySelectRequest;
-import spring.backend.activity.dto.response.UserActivitySelectResponse;
+import spring.backend.activity.presentation.dto.request.UserActivitySelectRequest;
+import spring.backend.activity.presentation.dto.response.UserActivitySelectResponse;
 import spring.backend.activity.exception.ActivityErrorCode;
 import spring.backend.core.exception.DomainException;
 import spring.backend.member.domain.entity.Member;
